+++ PhenomRaidTools/Source/Static.lua
@@ -1,0 +1,47 @@
+local PRT = LibStub("AceAddon-3.0"):GetAddon("PhenomRaidTools")
+
+-------------------------------------------------------------------------------
+-- Public API
+
+PRT.Static = {
+  Tables = {
+    RaidTargets = {
+      [0] = {
+        id = 0,
+        name = "None"
+      },
+      [1] = {
+        id = 1,
+        name = "Star |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_1:16|t"
+      },
+      [2] = {
+        id = 2,
+        name = "Circle |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_2:16|t"
+      },
+      [3] = {
+        id = 3,
+        name = "Diamond |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_3:16|t"
+      },
+      [4] = {
+        id = 4,
+        name = "Triangle |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_4:16|t"
+      },
+      [5] = {
+        id = 5,
+        name = "Moon |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_5:16|t"
+      },
+      [6] = {
+        id = 6,
+        name = "Square |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_6:16|t"
+      },
+      [7] = {
+        id = 7,
+        name = "Cross |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_7:16|t"
+      },
+      [8] = {
+        id = 8,
+        name = "Skull |TInterface\\TargetingFrame\\UI-RaidTargetingIcon_8:16|t"
+      }
+    }
+  }
+}